requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
httpx>=0.27
python-dotenv>=1.0
google-generativeai>=0.7.2


